+++ cmd/dbaas_user_show_kafka.go
@@ -1,0 +1,35 @@
+package cmd
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/exoscale/cli/pkg/globalstate"
+	"github.com/exoscale/cli/pkg/output"
+	v3 "github.com/exoscale/egoscale/v3"
+)
+
+func (c *dbaasUserShowCmd) showKafka(ctx context.Context) (output.Outputter, error) {
+
+	client, err := switchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(c.Zone))
+	if err != nil {
+		return &dbaasUserShowOutput{}, err
+	}
+
+	s, err := client.GetDBAASServiceKafka(ctx, c.Name)
+	if err != nil {
+		return &dbaasUserShowOutput{}, err
+	}
+
+	for _, u := range s.Users {
+		if u.Username == c.Username {
+			return &dbaasUserShowOutput{
+				Username: c.Username,
+				Type:     u.Type,
+			}, nil
+		}
+
+	}
+
+	return &dbaasUserShowOutput{}, fmt.Errorf("user %q not found for service %q", c.Username, c.Name)
+}